test: move test case in right file
